Move _infer_line_data back.
